chore: peer review
- Changed the name of the files
- added additional unit tests
